Cleanups and bug fix for password not being saved in DB
