+++ backend_app/.env.example
@@ -1,1 +1,1 @@
-DEV_DB_NAME=dbdev
+DEV_DB_NAME=rentalDevDB
