+++ backend_app/server/routes/Authentication.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const bcrypt = require("bcrypt");
-const { ajv } = require("../utils/validation/validation");
-const { createUser } = require("../utils/PersistenceUtils");
+const { ajv } = require("../validation/validation");
+const { createUser } = require("../../database/utils/PersistenceUtils");
 
 const SALT_ROUNDS = 10;
 const AuthenticationRouter = express.Router();
@@ -12,6 +12,7 @@ AuthenticationRouter.post("/signup", async (req, res, next) => {
 
   if (!validateUser(newUserData)) {
     next("400");
+    return;
   }
 
   try {
@@ -22,11 +23,13 @@ AuthenticationRouter.post("/signup", async (req, res, next) => {
 
     if (created === -1) {
       next("500");
+      return;
     }
 
     res.status(201).send("User successfully created");
   } catch (e) {
     next("500");
+    return;
   }
 });
