+++ backend_app/server/utils/ErrorHandler.js
@@ -3,19 +3,17 @@ const ErrorHandler = (err, _req, res, next) => {
   let status;
   let message;
 
-  switch (err.message) {
-  case "404":
+  if (err === "404") {
     status = 404;
     message = { message: "Ressources Not found" };
-    break;
-  case "400":
+  } else if (err === "400") {
     status = 400;
     message = { message: "Invalid Request" };
-    break;
-  default:
+  } else {
     status = 500;
     message = { message: "Internal Server Error" };
   }
+
   res.status(status).send(message);
 };
